records: merge dojson hep receivers
