+++ inspirehep/modules/records/receivers.py
@@ -26,12 +26,13 @@ from __future__ import absolute_import, division, print_function
 
 import six
 from flask import current_app
+from itertools import chain
 
 from invenio_indexer.signals import before_record_index
 from invenio_records.signals import before_record_insert, before_record_update
 
 from inspirehep.dojson.utils import classify_field, get_recid_from_ref
-from inspirehep.utils.date import create_valid_date
+from inspirehep.utils.date import create_earliest_date, create_valid_date
 from inspirehep.utils.helpers import force_force_list
 from inspirehep.utils.record_getter import get_db_record
 from inspirehep.utils.record import get_value, soft_delete_pidstore_for_record
@@ -43,6 +44,26 @@ from .signals import after_record_enhanced
 #
 # before_record_index
 #
+
+@before_record_index.connect
+def earliest_date(sender, json, *args, **kwargs):
+    """Find and assign the earliest date to a HEP paper."""
+    date_paths = [
+        'preprint_date',
+        'thesis.date',
+        'thesis.defense_date',
+        'publication_info.year',
+        'creation_modification_date.creation_date',
+        'imprints.date',
+    ]
+
+    dates = list(chain.from_iterable(
+        [force_force_list(get_value(json, path)) for path in date_paths]))
+
+    earliest_date = create_earliest_date(dates)
+    if earliest_date:
+        json['earliest_date'] = earliest_date
+
 
 @before_record_index.connect
 def enhance_record(sender, json, *args, **kwargs):
